requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
openai>=1.3.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
pandas>=2.1.0
numpy>=1.24.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
sqlalchemy>=2.0.0
schedule>=1.2.0
feedparser>=6.0.10
yfinance>=0.2.18
tweepy>=4.14.0
newspaper3k>=0.2.8
textblob>=0.17.1
scikit-learn>=1.3.0
nltk>=3.8.1
pyyaml>=6.0.1
lxml[html_clean]>=6.0.0
flask>=2.3.0
